Egg version bump. Moved tokens and activity hooks to being defined in Interpreters. Fixed some bugs in the setup function, and changed it so that it doesn't     have to be redefined in every subclass. All filter funcitons are in visioninterpreter.
 Changes to be committed:
	modified:   setup.py
	modified:   vision/visioninterpreter.py
	modified:   vision/visionparser.py
	modified:   vision/visionscanner.py
